Extract org.elixir_lang.beam.project_view_node.Module
